Move Publishers to active status on promo after KYC
